penumbra: regenerate protos
