+++ crates/proto/src/gen/penumbra.core.component.chain.v1alpha1.rs
@@ -34,22 +34,6 @@ impl ::prost::Name for Ratio {
         ::prost::alloc::format!("penumbra.core.component.chain.v1alpha1.{}", Self::NAME)
     }
 }
-/// Parameters for Fuzzy Message Detection
-#[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct FmdParameters {
-    #[prost(uint32, tag = "1")]
-    pub precision_bits: u32,
-    #[prost(uint64, tag = "2")]
-    pub as_of_block_height: u64,
-}
-impl ::prost::Name for FmdParameters {
-    const NAME: &'static str = "FmdParameters";
-    const PACKAGE: &'static str = "penumbra.core.component.chain.v1alpha1";
-    fn full_name() -> ::prost::alloc::string::String {
-        ::prost::alloc::format!("penumbra.core.component.chain.v1alpha1.{}", Self::NAME)
-    }
-}
 /// Chain-specific genesis content.
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
