+++ crates/proto/src/gen/penumbra.core.component.shielded_pool.v1alpha1.serde.rs
@@ -383,6 +383,121 @@ impl<'de> serde::Deserialize<'de> for EventSpend {
         deserializer.deserialize_struct("penumbra.core.component.shielded_pool.v1alpha1.EventSpend", FIELDS, GeneratedVisitor)
     }
 }
+impl serde::Serialize for FmdParameters {
+    #[allow(deprecated)]
+    fn serialize<S>(&self, serializer: S) -> std::result::Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        use serde::ser::SerializeStruct;
+        let mut len = 0;
+        if self.precision_bits != 0 {
+            len += 1;
+        }
+        if self.as_of_block_height != 0 {
+            len += 1;
+        }
+        let mut struct_ser = serializer.serialize_struct("penumbra.core.component.shielded_pool.v1alpha1.FmdParameters", len)?;
+        if self.precision_bits != 0 {
+            struct_ser.serialize_field("precisionBits", &self.precision_bits)?;
+        }
+        if self.as_of_block_height != 0 {
+            #[allow(clippy::needless_borrow)]
+            struct_ser.serialize_field("asOfBlockHeight", ToString::to_string(&self.as_of_block_height).as_str())?;
+        }
+        struct_ser.end()
+    }
+}
+impl<'de> serde::Deserialize<'de> for FmdParameters {
+    #[allow(deprecated)]
+    fn deserialize<D>(deserializer: D) -> std::result::Result<Self, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        const FIELDS: &[&str] = &[
+            "precision_bits",
+            "precisionBits",
+            "as_of_block_height",
+            "asOfBlockHeight",
+        ];
+
+        #[allow(clippy::enum_variant_names)]
+        enum GeneratedField {
+            PrecisionBits,
+            AsOfBlockHeight,
+        }
+        impl<'de> serde::Deserialize<'de> for GeneratedField {
+            fn deserialize<D>(deserializer: D) -> std::result::Result<GeneratedField, D::Error>
+            where
+                D: serde::Deserializer<'de>,
+            {
+                struct GeneratedVisitor;
+
+                impl<'de> serde::de::Visitor<'de> for GeneratedVisitor {
+                    type Value = GeneratedField;
+
+                    fn expecting(&self, formatter: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+                        write!(formatter, "expected one of: {:?}", &FIELDS)
+                    }
+
+                    #[allow(unused_variables)]
+                    fn visit_str<E>(self, value: &str) -> std::result::Result<GeneratedField, E>
+                    where
+                        E: serde::de::Error,
+                    {
+                        match value {
+                            "precisionBits" | "precision_bits" => Ok(GeneratedField::PrecisionBits),
+                            "asOfBlockHeight" | "as_of_block_height" => Ok(GeneratedField::AsOfBlockHeight),
+                            _ => Err(serde::de::Error::unknown_field(value, FIELDS)),
+                        }
+                    }
+                }
+                deserializer.deserialize_identifier(GeneratedVisitor)
+            }
+        }
+        struct GeneratedVisitor;
+        impl<'de> serde::de::Visitor<'de> for GeneratedVisitor {
+            type Value = FmdParameters;
+
+            fn expecting(&self, formatter: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+                formatter.write_str("struct penumbra.core.component.shielded_pool.v1alpha1.FmdParameters")
+            }
+
+            fn visit_map<V>(self, mut map_: V) -> std::result::Result<FmdParameters, V::Error>
+                where
+                    V: serde::de::MapAccess<'de>,
+            {
+                let mut precision_bits__ = None;
+                let mut as_of_block_height__ = None;
+                while let Some(k) = map_.next_key()? {
+                    match k {
+                        GeneratedField::PrecisionBits => {
+                            if precision_bits__.is_some() {
+                                return Err(serde::de::Error::duplicate_field("precisionBits"));
+                            }
+                            precision_bits__ = 
+                                Some(map_.next_value::<::pbjson::private::NumberDeserialize<_>>()?.0)
+                            ;
+                        }
+                        GeneratedField::AsOfBlockHeight => {
+                            if as_of_block_height__.is_some() {
+                                return Err(serde::de::Error::duplicate_field("asOfBlockHeight"));
+                            }
+                            as_of_block_height__ = 
+                                Some(map_.next_value::<::pbjson::private::NumberDeserialize<_>>()?.0)
+                            ;
+                        }
+                    }
+                }
+                Ok(FmdParameters {
+                    precision_bits: precision_bits__.unwrap_or_default(),
+                    as_of_block_height: as_of_block_height__.unwrap_or_default(),
+                })
+            }
+        }
+        deserializer.deserialize_struct("penumbra.core.component.shielded_pool.v1alpha1.FmdParameters", FIELDS, GeneratedVisitor)
+    }
+}
 impl serde::Serialize for GenesisContent {
     #[allow(deprecated)]
     fn serialize<S>(&self, serializer: S) -> std::result::Result<S::Ok, S::Error>
